Refactor and enhance ProteinMPNN codebase, introduced first pass for STE optimization
- Updated type annotations to improve clarity and maintainability, replacing enums with string literals for attention masks and decoding approaches.
- Refactored the sampling and optimization functions to utilize one-hot encoded sequences instead of integer indices.
- Introduced a new configuration class for sampling strategies, replacing the previous enum-based approach.
- Added comprehensive unit tests for the new sampling and optimization functionalities, ensuring robustness and correctness.
- Improved mock data generation in tests for better coverage and reliability.
- Removed deprecated enum tests and streamlined model version handling in tests.
- Enhanced the handling of model parameters and weights throughout the codebase for consistency.

+++ src/prxteinmpnn/utils/aa_convert.py
@@ -18,9 +18,9 @@ _MPNN_TO_AF_PERM = jnp.array(
 
 def af_to_mpnn(sequence: ProteinSequence) -> ProteinSequence:
   """Convert a sequence of integer indices from AlphaFold's to ProteinMPNN's alphabet order."""
-  return _AF_TO_MPNN_PERM[sequence]
+  return _AF_TO_MPNN_PERM[sequence].astype(jnp.int8)
 
 
 def mpnn_to_af(sequence: ProteinSequence) -> ProteinSequence:
   """Convert a sequence of integer indices from ProteinMPNN's to AlphaFold's alphabet order."""
-  return _MPNN_TO_AF_PERM[sequence]
+  return _MPNN_TO_AF_PERM[sequence].astype(jnp.int8)
